// src/types/api.d.ts
type Creator = {
  defaultAvatar: boolean
  province: number
  authStatus: number
  followed: boolean
  avatarUrl: string
  accountStatus: number
  gender: number
  city: number
  birthday: number
  userId: number
  userType: number
  nickname: string
  signature: string
  description: string
  detailDescription: string
  avatarImgId: number
  backgroundImgId: number
  backgroundUrl: string
  authority: number
  mutual: boolean
  expertTags: null
  experts: null
  djStatus: number
  vipType: number
  remarkName: null
  authenticationTypes: number
  avatarDetail: null
  avatarImgIdStr: string
  backgroundImgIdStr: string
  anchor: boolean
  avatarImgId_str: string
}
export type SongSheetType = {
  subscribers: any[]
  subscribed: boolean
  creator: Creator
  artists: null
  tracks: null
  top: boolean
  updateFrequency: null
  backgroundCoverId: number
  backgroundCoverUrl: null
  titleImage: number
  titleImageUrl: null
  englishTitle: null
  opRecommend: boolean
  recommendInfo: null
  subscribedCount: number
  cloudTrackCount: number
  userId: number
  totalDuration: number
  coverImgId: number
  privacy: number
  trackUpdateTime: number
  trackCount: number
  updateTime: number
  commentThreadId: string
  coverImgUrl: string
  specialType: number
  anonimous: boolean
  createTime: number
  highQuality: boolean
  newImported: boolean
  trackNumberUpdateTime: number
  playCount: number
  adType: number
  description: null
  tags: string[]
  ordered: boolean
  status: number
  name: string
  id: number
  coverImgId_str: string
  sharedUsers: null
  shareStatus: null
  copied: boolean
}
export type SongSheetList = SongSheetType[]